product repo&service

+++ WHManager.BusinessLogic/Models/Manufacturer.cs
@@ -8,6 +8,6 @@ namespace WHManager.BusinessLogic.Models
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int Nip { get; set; }
+        public double Nip { get; set; }
     }
 }
