+++ WHManager.BusinessLogic/Models/Product.cs
@@ -11,8 +11,8 @@ namespace WHManager.BusinessLogic.Models
         public ProductType Type { get; set; }
         public Tax Tax { get; set; }
         public Manufacturer Manufacturer { get; set; }
-        public double PriceBuy { get; set; }
-        public double PriceSell { get; set; }
+        public decimal PriceBuy { get; set; }
+        public decimal PriceSell { get; set; }
         public bool InStock { get; set; }
     }
 }
