+++ WHManager.DataAccess/Models/Product.cs
@@ -12,8 +12,8 @@ namespace WHManager.DataAccess.Models
         public ProductType Type { get; set; }
         public Tax Tax { get; set; }
         public Manufacturer Manufacturer { get; set; }
-        public double PriceBuy { get; set; }
-        public double PriceSell { get; set; }
+        public decimal PriceBuy { get; set; }
+        public decimal PriceSell { get; set; }
         public bool InStock { get; set; }
     }
 }
