+++ WHManager.DataAccess/Repositories/Interfaces/IManufacturerRepository.cs
@@ -6,12 +6,12 @@ namespace WHManager.DataAccess.Repositories
 {
     public interface IManufacturerRepository
     {
-        Task<Manufacturer> AddManufacturerAsync(int id, string name, int nip);
+        Task<Manufacturer> AddManufacturerAsync(int id, string name, double nip);
         IEnumerable<Manufacturer> GetManufacturers();
         Manufacturer GetManufacturer(int id);
         Task DeleteManufacturerAsync(int id);
-        Task UpdateManufacturerAsync(int id, string name, int nip);
-        Manufacturer GetManufacturerByNip(int nip);
+        Task UpdateManufacturerAsync(int id, string name, double nip);
+        Manufacturer GetManufacturerByNip(double nip);
         IEnumerable<Manufacturer> GetManufacturersByName(string name);
     }
 }
